CASMPET-6437 add initial testing for setting image in BSS
